null file is not produces 500 error
