+++ MangoAPI.BusinessLogic/ApiCommands/Users/RegisterCommand.cs
@@ -1,5 +1,4 @@
 ﻿using MangoAPI.BusinessLogic.Responses;
-using MangoAPI.Domain.Enums;
 using MediatR;
 
 namespace MangoAPI.BusinessLogic.ApiCommands.Users
